[#3] update dependency-versions to match CRAN
A number of the package versions in the meta.yaml for r-diagrammer preceded
those in the 'imports' section of the CRAN-DiagrammeR page; this despite
the current version of DiagrammeR on CRAN being the same as the version on
conda-forge (v-1.0.0).

The dependency versions for r-diagrammer=1.0.0 were updated to match those
at CRAN for DiagrammeR 1.0.0.

build/number was increased by 1 since the r-diagrammer package version was
not changed.